info-streaming/requirements.txt
httpx~=0.27.0
redis~=5.0.4
fastapi[standard]>=0.115.6
uvicorn>=0.32.1
websocket-client>=0.59.0